Move testConnection tests to separate class
