+++ tests/Keboola/MongoDbExtractor/ApplicationTest.php
@@ -27,137 +27,6 @@ class ApplicationTest extends \PHPUnit_Framework_TestCase
         $this->fs->remove($this->path);
     }
 
-    public function testActionTestConnectionOk()
-    {
-        $json = <<<JSON
-{
-  "parameters": {
-    "db": {
-      "host": "mongodb",
-      "port": 27017,
-      "database": "test"
-    },
-    "exports": [
-      {
-        "name": "bakeries",
-        "id": 123,
-        "collection": "restaurants",
-        "incremental": true,
-        "mapping": {
-          "_id.\$oid": "id"
-        }
-      }
-    ]
-  }
-}
-JSON;
-        $config = (new JsonDecode(true))->decode($json, JsonEncoder::FORMAT);
-
-        $application = new Application($config);
-        $application->actionTestConnection();
-
-        $this->assertSame(['status' => 'ok'], $application->actionTestConnection());
-    }
-
-    public function testActionTestConnectionFailWrongHost()
-    {
-        $this->expectException(\MongoDB\Driver\Exception\ConnectionTimeoutException::class);
-
-        $json = <<<JSON
-{
-  "parameters": {
-    "db": {
-      "host": "locahost",
-      "port": 27017,
-      "database": "test"
-    },
-    "exports": [
-      {
-        "name": "bakeries",
-        "id": 123,
-        "collection": "restaurants",
-        "incremental": true,
-        "mapping": {
-          "_id.\$oid": "id"
-        }
-      }
-    ]
-  }
-}
-JSON;
-        $config = (new JsonDecode(true))->decode($json, JsonEncoder::FORMAT);
-
-        $application = new Application($config);
-        $application->actionTestConnection();
-    }
-
-    public function testActionTestConnectionFailWithoutPassword()
-    {
-        $this->expectException(\MongoDB\Driver\Exception\RuntimeException::class);
-        $this->expectExceptionMessageRegExp('~not authorized~');
-
-        $json = <<<JSON
-{
-  "parameters": {
-    "db": {
-      "host": "mongodb-auth",
-      "port": 27017,
-      "database": "test"
-    },
-    "exports": [
-      {
-        "name": "bakeries",
-        "id": 123,
-        "collection": "restaurants",
-        "incremental": true,
-        "mapping": {
-          "_id.\$oid": "id"
-        }
-      }
-    ]
-  }
-}
-JSON;
-        $config = (new JsonDecode(true))->decode($json, JsonEncoder::FORMAT);
-
-        $application = new Application($config);
-        $application->actionTestConnection();
-    }
-
-    public function testActionTestConnectionFailWrongPassword()
-    {
-        $this->expectException(\MongoDB\Driver\Exception\AuthenticationException::class);
-
-        $json = <<<JSON
-{
-  "parameters": {
-    "db": {
-      "host": "mongodb-auth",
-      "port": 27017,
-      "database": "test",
-      "user": "user",
-      "password": "random-password"
-    },
-    "exports": [
-      {
-        "name": "bakeries",
-        "id": 123,
-        "collection": "restaurants",
-        "incremental": true,
-        "mapping": {
-          "_id.\$oid": "id"
-        }
-      }
-    ]
-  }
-}
-JSON;
-        $config = (new JsonDecode(true))->decode($json, JsonEncoder::FORMAT);
-
-        $application = new Application($config);
-        $application->actionTestConnection();
-    }
-
     public function testActionRunFull()
     {
         $json = <<<JSON
