+++ tests/Keboola/MongoDbExtractor/ApplicationTestConnectionTest.php
@@ -1,0 +1,164 @@
+<?php
+
+namespace Keboola\MongoDbExtractor;
+
+use MongoDB\Driver\Exception\AuthenticationException;
+use MongoDB\Driver\Exception\RuntimeException;
+use MongoDB\Driver\Exception\ConnectionTimeoutException;
+use Symfony\Component\Serializer\Encoder\JsonDecode;
+use Symfony\Component\Serializer\Encoder\JsonEncoder;
+use Symfony\Component\Filesystem\Filesystem;
+
+class ApplicationTestConnectionTest extends \PHPUnit_Framework_TestCase
+{
+    /** @var Filesystem */
+    private $fs;
+
+    protected $path = '/tmp/application-test-connection-test';
+
+    protected function setUp()
+    {
+        $this->fs = new Filesystem;
+        $this->fs->remove($this->path);
+        $this->fs->mkdir($this->path);
+
+        parent::setUp();
+    }
+
+    protected function tearDown()
+    {
+        $this->fs->remove($this->path);
+    }
+
+    public function testActionTestConnectionOk()
+    {
+        $json = <<<JSON
+{
+  "parameters": {
+    "db": {
+      "host": "mongodb",
+      "port": 27017,
+      "database": "test"
+    },
+    "exports": [
+      {
+        "name": "bakeries",
+        "id": 123,
+        "collection": "restaurants",
+        "incremental": true,
+        "mapping": {
+          "_id.\$oid": "id"
+        }
+      }
+    ]
+  }
+}
+JSON;
+        $config = (new JsonDecode(true))->decode($json, JsonEncoder::FORMAT);
+
+        $application = new Application($config);
+        $application->actionTestConnection();
+
+        $this->assertSame(['status' => 'ok'], $application->actionTestConnection());
+    }
+
+    public function testActionTestConnectionFailWrongHost()
+    {
+        $this->expectException(ConnectionTimeoutException::class);
+
+        $json = <<<JSON
+{
+  "parameters": {
+    "db": {
+      "host": "locahost",
+      "port": 27017,
+      "database": "test"
+    },
+    "exports": [
+      {
+        "name": "bakeries",
+        "id": 123,
+        "collection": "restaurants",
+        "incremental": true,
+        "mapping": {
+          "_id.\$oid": "id"
+        }
+      }
+    ]
+  }
+}
+JSON;
+        $config = (new JsonDecode(true))->decode($json, JsonEncoder::FORMAT);
+
+        $application = new Application($config);
+        $application->actionTestConnection();
+    }
+
+    public function testActionTestConnectionFailWithoutPassword()
+    {
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessageRegExp('~not authorized~');
+
+        $json = <<<JSON
+{
+  "parameters": {
+    "db": {
+      "host": "mongodb-auth",
+      "port": 27017,
+      "database": "test"
+    },
+    "exports": [
+      {
+        "name": "bakeries",
+        "id": 123,
+        "collection": "restaurants",
+        "incremental": true,
+        "mapping": {
+          "_id.\$oid": "id"
+        }
+      }
+    ]
+  }
+}
+JSON;
+        $config = (new JsonDecode(true))->decode($json, JsonEncoder::FORMAT);
+
+        $application = new Application($config);
+        $application->actionTestConnection();
+    }
+
+    public function testActionTestConnectionFailWrongPassword()
+    {
+        $this->expectException(AuthenticationException::class);
+        $this->expectExceptionMessageRegExp('~Authentication failed~');
+
+        $json = <<<JSON
+{
+  "parameters": {
+    "db": {
+      "host": "mongodb-auth",
+      "port": 27017,
+      "database": "test",
+      "user": "user",
+      "password": "random-password"
+    },
+    "exports": [
+      {
+        "name": "bakeries",
+        "id": 123,
+        "collection": "restaurants",
+        "incremental": true,
+        "mapping": {
+          "_id.\$oid": "id"
+        }
+      }
+    ]
+  }
+}
+JSON;
+        $config = (new JsonDecode(true))->decode($json, JsonEncoder::FORMAT);
+
+        $application = new Application($config);
+        $application->actionTestConnection();
+    }
+}
